backend/requirements.txt
asgiref==3.3.1
Django==3.1.4
djangorestframework==3.12.2
pytz==2020.4
sqlparse==0.4.1
psycopg2>=2.8
django-cors-headers==3.6.0
djangorestframework-simplejwt==4.6.0
gunicorn==20.0.4
celery==5.0.5
redis==3.5.3
docx2python==1.27.1